backend/pyproject.toml
[project]
name = "geo-backend"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "fastapi",
  "uvicorn",
  "sqlalchemy",
  "psycopg[binary]",
  "pydantic",
  "pydantic-settings",
  "python-jose[cryptography]",
  "httpx",
  "pytest"
]

[tool.pytest.ini_options]
pythonpath = ["."]
